Move iptvsimple update code
