+++ addon.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="plugin.program.iptv-list_refresh" name="IPTV-list Refresh" version="0.0.5" provider-name="zjoasan">
+<addon id="plugin.program.iptv-list_refresh" name="IPTV-list Refresh" version="0.0.7" provider-name="zjoasan">
 	<requires>
 		<import addon="xbmc.python" version="3.0.0"/>
 	</requires>
